♻️ Entity field modify

+++ src/main/java/com/t2m/g2nee/shop/couponset/CouponType/domain/CouponType.java
@@ -23,7 +23,7 @@ public class CouponType {
     private Integer period;
     @Enumerated(EnumType.STRING)
     private Type type;
-    private Long discount;
+    private BigDecimal discount;
     private BigDecimal minimumOrderAmount;
     private BigDecimal maximumDiscount;
